==> 03 ELI vs NIT.Rmd <==
---
title: "Análisis de expresión diferencial ELI vs NIT"
output:
  html_notebook: default
bibliography: bibliografía/bibliography.bib
---

# librerias usadas

```{r}
library(tidyverse); library(DESeq2); library(apeglm); library(Cairo); library(pheatmap); library(EnhancedVolcano); library(clusterProfiler); library(enrichplot); library(pathview)
library(AnnotationDbi); library(org.Hs.eg.db); library(pheatmap) # el paquete de anotaciones de humanos usando Entrez
```

# Preparación del análisis de expresión de ELI vs NIT

A partir de los datos normalizados con DESeq2 se extraen los resultados de las comparaciones dos a dos, para hacer las comparaciones se pone en el numerador el mayor grado de infiltración (ELI>SFI>NIT), de forma que cuando los genes sobreexpresados serán aquellos que se expresen mas en la muestra con mayor infiltración, en este comparación, los genes sobreexpresados serán los más transcritos en las muestras ELI.

## Filtrado

Se filtran los genes por el fold change (parámetro lfcThreshold) y por el valor de significación (el parámentro alpha). El filtrado por defecto acepta como genes diferencialmente expresados cualquiera que tenga un p-valor menor de 0.1:

```{r}
load("materiales intermedios/dds_raw.Rda")

eli_nit <- results(dds, contrast=c("Group", "ELI", "NIT"))

summary(eli_nit) 

EnhancedVolcano(eli_nit, lab = "", x = 'log2FoldChange', FCcutoff = 0,
                pCutoff = 0.1, y = 'pvalue', title = "Volcano ELI vs NIT", 
                subtitle = paste0("Genes diferencialmente expresados: ", sum(eli_nit$padj < 0.1, na.rm=TRUE)),
                gridlines.major = F, gridlines.minor = F)
```

Según Conesa y colaboradores [@Conesa2016] para 10 replicas por grupos y con un nivel de significación del 0.05, el fold change de 1,25 tiene una potencia (probabilidad de detectar expresión diferencial) del 44%, mientras que con 1.5 FC la potencia sube al 91%, por lo que usamos FC=1.5. 

```{r}
FC <- 1.5; alpha <- 0.05
eli_nit <- results(dds, contrast=c("Group", "ELI", "NIT"), alpha = alpha)

summary(eli_nit)

EnhancedVolcano(eli_nit, FCcutoff = FC, lab = "", x = 'log2FoldChange',
                pCutoff = alpha, y = 'pvalue', title = "Volcano ELI vs NIT", 
                subtitle = paste0("Genes diferencialmente expresados: ",
                                  sum((eli_nit$padj < alpha)& abs(eli_nit$log2FoldChange)>FC, na.rm=T)),
                gridlines.major = F, gridlines.minor = F, )
```

## Shrink

Encoger (Shrink) el fold change de los genes con pocas cuentas reduce el ruido, es util en rankings y visualizacion, para evitar filtros extra, que son arbitrarios. En el paquete DESEq hay tres tipos de Shrink, paso a compararlos para elegir el mejor.

```{r message=F}
eli_nit_shr <- lfcShrink(dds, coef="Group_ELI_vs_NIT", type= "apeglm", res=eli_nit) 
eli_nit_ashr <- lfcShrink(dds, contrast=c("Group", "ELI", "NIT"), type = "ashr", res=eli_nit) # Con ashr shrink

par(mfrow=c(1,3))
  plotMA(eli_nit, ylim = c(-5,10), main = "no shrink")
  plotMA(eli_nit_shr, ylim = c(-5,10), main = "apeglm") 
  plotMA(eli_nit_ashr, ylim = c(-5,10), main = "ashr")
```

El shrinkage normal es peor que los otros dos métodos, muy similares entre sí, voy a usar el método ashr porque apeglm solo funciona con coef, no con contrast; y con coef no puedo hacer la comparación ELI vs SFI.

```{r}
summary(eli_nit_ashr)

(p <- EnhancedVolcano(eli_nit_ashr, FCcutoff = FC, lab = "", x = 'log2FoldChange',
                pCutoff = alpha, y = 'pvalue', title = "Volcano ELI vs NIT", 
                subtitle = paste0("Genes diferencialmente expresados: ",
                                  sum((eli_nit_ashr$padj < alpha)& abs(eli_nit_ashr$log2FoldChange)>FC, na.rm=T)),
                gridlines.major = F, gridlines.minor = F))
dir.create("resultados/ElI vs NIT")
dir.create("resultados/ElI vs NIT/imagenes")

cairo_pdf("resultados/ElI vs NIT/Volcano.pdf")
p
dev.off()

png("resultados/ElI vs NIT/imagenes/Volcano.png", height = 1500, width = 2000, res=200)
p
dev.off()
```
Los genes elegidos son los marcados en rojo, son `r sum((eli_nit_ashr$padj < alpha)& abs(eli_nit_ashr$log2FoldChange)>FC, na.rm=T)` de un total de `r nrow(eli_nit_ashr)`, lo que supone el `r round(sum((eli_nit_ashr$padj < alpha)& abs(eli_nit_ashr$log2FoldChange)>FC, na.rm=T)/(nrow(eli_nit_ashr)),4)*100`%.

Distribución de p-valores. Se usa como control de calidad, la forma que tiene indica una buena calidad que muchos de los genes están diferencialmente expresados.

```{r}
# se quitan los genes cuyas cuentas normalizadas en promedio no llegan a 1 porque añaden ruido
pvalores_eli_nit_ashr <- eli_nit_ashr$pvalue[eli_nit_ashr$baseMean > 1] 

hist(pvalores_eli_nit_ashr, breaks = 0:20/20,
     col = "grey50", border = "white", probability = T,
     main = "apeglm shrink")
```

# Anotación y análisis de expresión

Usando el fold change de `r FC` y el nivel de significación `r alpha` los genes diferencialmente expresados son los siguientes:

```{r}
# corta los nombres a strings de 15, para dejar el nombre de ensembl y que así se pueda localizar con 
# mapIds con las keys
ens.str <- substr(rownames(eli_nit), 1, 15) 

columns(org.Hs.eg.db)

eli_nit_ashr$SYMBOL <- mapIds(org.Hs.eg.db, keys=ens.str, keytype="ENSEMBL", multiVals="first",
                              column="SYMBOL") # la columna que añade
eli_nit_ashr$ENTREZID <- mapIds(org.Hs.eg.db, keys=ens.str, keytype="ENSEMBL", multiVals="first",
                                column="ENTREZID")
eli_nit_ashr$GO <- mapIds(org.Hs.eg.db, keys=ens.str, keytype="ENSEMBL", multiVals="first",
                          column="GO")
eli_nit_ashr$PATH <- mapIds(org.Hs.eg.db, keys=ens.str, keytype="ENSEMBL", multiVals="first",
                            column="PATH")
eli_nit_ashr$UNIPROT <- mapIds(org.Hs.eg.db, keys=ens.str, keytype="ENSEMBL", multiVals="first",
                               column="UNIPROT")
eli_nit_ashr$ENZYME <- mapIds(org.Hs.eg.db, keys=ens.str, keytype="ENSEMBL", multiVals="first",
                              column="ENZYME")
```

Ordeno por p-valor y le doy formato de dataframe

```{r}
eli_nit_DGE <- as.data.frame(eli_nit_ashr[order(eli_nit_ashr$pvalue),]) 
# tabla con todos los genes
write.csv(as.data.frame(eli_nit_DGE), file = "resultados/ElI vs NIT/DGE.csv")
```

Hago una tabla aparte con los `r sum((eli_nit_ashr$padj < alpha)& abs(eli_nit_ashr$log2FoldChange)>FC, na.rm=T)` genes de expresión diferencial, esta es la tabla que se muestra en el informe.

```{r}
eli_nitSig <- as.data.frame(subset(eli_nit_ashr, padj < alpha & abs(log2FoldChange)>FC)) %>% arrange(-log2FoldChange)
# genes más sobreexpresados
head(eli_nitSig)
# genes más subexpresados
tail(eli_nitSig)

write.csv(as.data.frame(eli_nitSig), file = "resultados/ElI vs NIT/DGE_sig.csv")
```

# Análisis de sobrerrepresentación (ORA)

Consiste en un resumen de los procesos en los que más genes de los sobreexpresados y subexpresados intervienen. A continuación se muestran tanto las rutas metabólicas de KEGG como por categorías de GO más representadas. Fuente: https://yulab-smu.github.io/clusterProfiler-book/index.html

```{r}
# separo los genes
upregulated <- filter(eli_nitSig, log2FoldChange>FC) 
downregulated <- filter(eli_nitSig, log2FoldChange<(-FC)) 
# Quito los NAs de EntrezID
upregulated <- upregulated[!is.na(upregulated$ENTREZID),]
downregulated <- downregulated[!is.na(downregulated$ENTREZID),]

nrow(upregulated)
nrow(downregulated)
```

Al eliminar los genes cuyos ENTREZID no se ha localizado se reduce el número de genes disponible, por eso upregulted + downregulated no suma lo mismo que los genes diferencialmente expresados. Si se hace la suma con los vectores contanto los NAs, coincide.

## GO

```{r}
upregulated_GO <- enrichGO(upregulated$ENTREZID, 
                             OrgDb="org.Hs.eg.db", 
                             pvalueCutoff=0.05,
                             qvalueCutoff=0.05, 
                             ont = "BP")

upregulated_GO <- setReadable(upregulated_GO, "org.Hs.eg.db", keyType = "ENTREZID")

head(upregulated_GO)
```

```{r}
downregulated_GO <- enrichGO(downregulated$ENTREZID, 
                             OrgDb="org.Hs.eg.db", 
                             pvalueCutoff=0.05,
                             qvalueCutoff=0.05, 
                             ont = "BP")

downregulated_GO <- setReadable(downregulated_GO, "org.Hs.eg.db", keyType = "ENTREZID")

head(downregulated_GO)
```
### Términos GO más representados:

```{r}
p1 <- barplot(upregulated_GO, showCategory = 20) + ggtitle("Categorias GO sobreexpresadas")
p2 <- barplot(downregulated_GO, showCategory=20) + ggtitle("Categorias GO subexpresadas")

cairo_pdf("resultados/ElI vs NIT/GO barplot.pdf", height = 10, width = 20)
  plot_grid(p1, p2, ncol=2)
dev.off()

png("resultados/ElI vs NIT/imagenes/GO barplot.png", height = 1500, width = 2000, res=200)
  p1 # los subexpresados no tienen nada especialmente interesante
dev.off()
```

<!-- Lo mismo pero con puntos  -->

```{r}
# p1 <- dotplot(upregulated_GO, showCategory=20) + ggtitle("Categorias GO sobreexpresadas")
# p2 <- dotplot(downregulated_GO, showCategory=20) + ggtitle("Categorias GO subexpresadas")
# 
# cairo_pdf("resultados/ELIvsNIT GO dotplot.pdf", height = 10, width = 20)
#   plot_grid(p1, p2, ncol=2)
# dev.off()
```

### Redes de términos GO:

```{r}
cairo_pdf("resultados/ElI vs NIT/upregulated GO network.pdf", height = 10, width = 20)
  cnetplot(upregulated_GO, foldChange=upregulated$log2FoldChange)
dev.off()

png("resultados/ElI vs NIT/imagenes/upregulated GO network.png", height = 1500, width = 3000, res=200)
  cnetplot(upregulated_GO, foldChange=upregulated$log2FoldChange)
dev.off()

cairo_pdf("resultados/ElI vs NIT/downregulated GO network.pdf", height = 10, width = 20)
  cnetplot(downregulated_GO, foldChange=upregulated$log2FoldChange)
dev.off()

png("resultados/ElI vs NIT/imagenes/downregulated GO network.png", height = 1500, width = 3000, res=200)
  cnetplot(downregulated_GO, foldChange=upregulated$log2FoldChange)
dev.off()
```

## KEGG

```{r}
upregulated_KEGG <- enrichKEGG(gene = upregulated$ENTREZID,
                               organism = 'hsa', 
                               pvalueCutoff = 0.05)

upregulated_KEGG <- setReadable(upregulated_KEGG, "org.Hs.eg.db", keyType = "ENTREZID")

head(upregulated_KEGG)
```

```{r}
downregulated_KEGG <- enrichKEGG(gene = downregulated$ENTREZID,
                               organism = 'hsa', 
                               pvalueCutoff = 0.05)

downregulated_KEGG <- setReadable(downregulated_KEGG, "org.Hs.eg.db", keyType = "ENTREZID")

head(downregulated_KEGG)
```
No encuentra ninguna ruta sobrerrepresentada en los NIT, me centro en las muestras ELI.

### Rutas KEGG más representadas:

```{r}
p1 <- barplot(upregulated_KEGG, showCategory = 20) + ggtitle("Rutas KEGG sobreexpresadas")

cairo_pdf("resultados/ElI vs NIT/KEGG barplot.pdf", height = 10, width = 10)
  p1
dev.off()

png("resultados/ElI vs NIT/imagenes/KEGG barplot.png", height = 1500, width = 2000, res=200)
  p1
dev.off()

```

### Redes de rutas KEGG:

```{r}
cairo_pdf("resultados/ElI vs NIT/upregulated KEGG network.pdf", height = 10, width = 20)
    cnetplot(upregulated_KEGG, foldChange=upregulated$log2FoldChange)
dev.off()

png("resultados/ElI vs NIT/imagenes/upregulated KEGG network.png", height = 1500, width = 2000, res=200)
  cnetplot(upregulated_KEGG, foldChange=upregulated$log2FoldChange)
dev.off()
```

```{r}
library(beepr)
beep(3)
```
